feat: Add urn to collection and items
