+++ src/Collection/utils.ts
@@ -1,10 +1,7 @@
 import { getCurrentNetworkURNProtocol } from '../ethereum/utils'
-import { CollectionAttributes } from './Collection.types'
 
 export function getDecentralandCollectionURN(
-  collection: CollectionAttributes
+  collectionAddress: string
 ): string {
-  return `urn:decentraland:${getCurrentNetworkURNProtocol()}:collections-v2:${
-    collection.contract_address
-  }}`
+  return `urn:decentraland:${getCurrentNetworkURNProtocol()}:collections-v2:${collectionAddress}`
 }
