+++ src/Item/Item.types.ts
@@ -1,5 +1,5 @@
-import { MetricsAttributes, metricsSchema } from '../Metrics'
-import { WearableData, wearableSchema } from './wearable/types'
+import { MetricsAttributes } from '../Metrics'
+import { WearableData } from './wearable/types'
 
 export enum ItemType {
   WEARABLE = 'wearable',
@@ -38,48 +38,3 @@ export type ItemAttributes = {
   created_at: Date
   updated_at: Date
 }
-
-export const itemSchema = Object.freeze({
-  type: 'object',
-  properties: {
-    id: { type: 'string', format: 'uuid' },
-    urn: { type: 'string' },
-    name: { type: 'string', maxLength: 32 },
-    description: { type: ['string', 'null'], maxLength: 64 },
-    thumbnail: { type: 'string' },
-    eth_address: { type: 'string' },
-    collection_id: { type: ['string', 'null'], format: 'uuid' },
-    blockchain_item_id: { type: ['string', 'null'] },
-    price: { type: ['string', 'null'] },
-    beneficiary: { type: ['string', 'null'] },
-    rarity: {
-      type: ['string', 'null'],
-      enum: [...Object.values(ItemRarity), null],
-    },
-    total_supply: { type: 'number', minimum: 0 },
-    is_published: { type: 'boolean' },
-    is_approved: { type: 'boolean' },
-    type: { enum: Object.values(ItemType) },
-    data: { type: 'object', oneOf: [wearableSchema] },
-    metrics: metricsSchema,
-    contents: {
-      type: 'object',
-      additionalProperties: true,
-    },
-    created_at: { type: 'string' },
-    updated_at: { type: 'string' },
-  },
-  additionalProperties: false,
-  required: [
-    'id',
-    'name',
-    'description',
-    'eth_address',
-    'data',
-    'type',
-    'metrics',
-    'contents',
-    'created_at',
-    'updated_at',
-  ],
-})
