+++ src/Item/utils.ts
@@ -1,10 +1,11 @@
-import { CollectionAttributes } from '../Collection/Collection.types'
 import { getDecentralandCollectionURN } from '../Collection/utils'
 import { ItemAttributes } from './Item.types'
 
 export function getDecentralandItemURN(
   item: ItemAttributes,
-  collection: CollectionAttributes
+  collectionAddress: string
 ): string {
-  return `${getDecentralandCollectionURN(collection)}:${item.id}`
+  return `${getDecentralandCollectionURN(collectionAddress)}:${
+    item.blockchain_item_id
+  }`
 }
