+++ src/ethereum/api/Bridge.ts
@@ -1,11 +1,11 @@
 import { CollectionAttributes, Collection } from '../../Collection'
 import { ItemAttributes, Item, ItemRarity } from '../../Item'
 import { MetricsAttributes } from '../../Metrics'
+import { fromUnixTimestamp } from '../../utils/parse'
+import { WearableCategory, WearableData } from '../../Item/wearable/types'
 import { ItemFragment, CollectionFragment } from './fragments'
 import { collectionAPI } from './collection'
 import { Wearable } from './peer'
-import { fromUnixTimestamp } from '../../utils/parse'
-import { WearableCategory, WearableData } from '../../Item/wearable/types'
 
 export class Bridge {
   static async consolidateCollections(
@@ -147,6 +147,7 @@ export class Bridge {
     let contents: Record<string, string>
     let metrics: MetricsAttributes
     let in_catalyst: boolean
+    let urn: string | null = null
 
     if (catalystItem) {
       data = catalystItem.data
@@ -158,6 +159,12 @@ export class Bridge {
       contents = dbItem.contents
       metrics = dbItem.metrics
       in_catalyst = false
+    }
+
+    if (catalystItem) {
+      urn = catalystItem.id
+    } else if (remoteItem && remoteItem.urn) {
+      urn = remoteItem.urn
     }
 
     if (wearable) {
@@ -182,6 +189,7 @@ export class Bridge {
     return {
       ...dbItem,
       name,
+      urn,
       description,
       rarity,
       price: remoteItem.price,
